move SpoonMetaModel to spoon.test.metamodel
