+++ src/test/java/spoon/test/metamodel/MMContainerType.java
@@ -14,7 +14,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.metamodel;
+package spoon.test.metamodel;
 
 import java.util.List;
 import java.util.Map;
