+++ src/test/java/spoon/test/metamodel/MMField.java
@@ -14,7 +14,10 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.metamodel;
+package spoon.test.metamodel;
+
+import static spoon.test.metamodel.SpoonMetaModel.addUniqueObject;
+import static spoon.test.metamodel.SpoonMetaModel.getOrCreate;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -32,9 +35,6 @@ import spoon.reflect.path.CtRole;
 import spoon.reflect.reference.CtTypeParameterReference;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.support.DerivedProperty;
-
-import static spoon.metamodel.SpoonMetaModel.addUniqueObject;
-import static spoon.metamodel.SpoonMetaModel.getOrCreate;
 
 /**
  * Represents a field of Spoon model type.
